requirements.txt
python-telegram-bot
httpx
python-dotenv
requests
